Cart: Allow to add things in front or after the current translation.
Within the post view you can add the current one with some target
languages to the cart.
But those posts might be related to some other data
and loose their connections if the other ones were not translated.
To solve this a filter before
and after adding things to the cart is fired.
In that way some validation can be done,
parent posts,
terms or other related data can be added to the cart too.

TLM-67