Add transient notices to able to give users feedback about api calls
Previously the actions thrown ActionExceptions but since the handle method
is hooked in WordPress there is not way by our side to catch those,
so instead I created a transient notice that store the errors from the actions
and from the api calls into a transient and show them on page load.

This is useful even for when we need to add a notice before a page redirect.

The Api `create` methods now thrown a `ApiException` that we can catch to
create the notices for the users instead of returning a simple `null` value.

Also, created a new class `Api/Response` that collect all of the api server
responses and change the text a bit to be more understandable for the users.