Fix: To get the current blog language use `get_current_blog_id()`.
The old `get_current_site()` will return the ID of the current network.

See : https://codex.wordpress.org/WPMU_Functions/get_current_site
For more infos.

get_current_site() was inherited from MU, and is named based on the old
terminology which spoke of multiple "blogs" on a "site".

We now speak of multiple "sites" on a "network" instead,
but some of the old terminology still lives on in some function names.
This function returns information about the current network.
To get information about the current site on the network,
see get_current_blog_id() and get_blog_details().