TM-127: Move About page under Settings Tab
The style.css is renamed to `settings.scss` and every scss file should be
stored under the same directory.

All of the compiled css start by now will are enqueued from `assets/css`.
This to allow us to remove the scss files if we don't want to provide them
in the package release.

The tabs content are now moved on separated files under `views/options-page/tabs`
for maintainability instead of having all of the markup tabs in one place.

Introduced `scss-lint.yml`.