TM-151: Introduce Network Settings
Allow to define a network settings as fallback if the settings are not found
into the specific site.

The options for the network are handled by `SiteOptionsHandler`.

Introducing `ApiSettings` a class to retrieve the api settings based on context
that also implements the fallback to the network options.

This class get the constants from the `OptionSettings` for URL and TOKEN.

So now if you want to retrieve the api settings just call:

`ApiSettings::token()` to retrieve the token (contextualized)
`ApiSettings::url()` to retrieve the url (contextualized)

In case you want to retrieve the option uncontextualized (means using the fallback)
just pass `false` as the first parameter to the functions.

The function `translationmanager_api()` retrieve the settings always uncontextualized,
this because we use it across the plugin and not to retrieve the settings for specific
settings pages.

The `Pages` are got a new method `url` that is a static method useful to retrieve
the url page based on the context.

The submit button changed the `name` attribute to `translationmanager_save_settings`
so we know we are handle the plugin settings and not other settings.

This should be used as general name for all settings.

Introduced `nonce.php` input type template, so we can reuse the same markup.