TM-151: Fix: Be sure the plugin is network activate to prevent strange behaviors.
It's possible that the plugin was activated, settings were stored and then
the plugin was deactivated, we can have unexpected options value because the plugin
still get the information from the network instead of the current site.

Introduce a new wrapper function `is_plugin_active_for_network` to make
tests during the plugin bootstrap.

Also fixed the pages url when multisite is active, always check if the plugin
is network activated.

And introduced a new Plugin function named `plugin_file` that retrieve
the base plugin file path.