Refactor: Always use the project term ID instead of the slug to make the logic coherent across the project.
Only use the slug when call `wp_set_post_terms` because the project taxonomy
isn't a hierarchical one. So we need to pass the term slug instead of and ID.

Also, changed all references to `_translationmanager_project_id` by removing
the underscore prefix.

In the old code the version prefixed with the underscore was used to handle it
as term slug instead of ID.

And, in the taxonomy edit page we previously used the query arg `translationmanager_project`
now this has been changed to be `translationmanager_project_id`.

So now we deal always with term ID and always with the same query argument key.