TM-172: Introduce support for MLP3 in Container/Service context.
The classes `Mlp_Site_Relations` and `Mlp_Content_Relations` are now created by the `Adapter`
within the Mlp module package.

This to give support for Mlp2 and version 3.

There is a mapper inside the Adapter class that call the correct method for the instances
depending on the version of the plugin.

When you need to add a new method you must include a new map record.

Also:

- Fix Missed to update the `Project` to `Taxonomy` class for `get_project_link` method call.
  This after the refactor that switched the dir and class name for make the plugin more DDD.
  See 78a9732
- Fix `get_languages_by_site_id` must use filter and not directly implement code for mlp.
- Fix `bulk_translate_projects_by_request_posts` cast to int the project_id,
  bad idea because the `-1` is used to know if a new project must be created.
  Keep it as string.