Project Update loop through translations twice but import only first translation
# Project Update loop twice

Because of data struct retrieved by the api call we'll loop through the
translations twice.

The `$translation_data['items']` include for every entry two elements:
one is an array of meta, the other item is the status of the project.

TM is looping unnecessary both.

# Translations import only the first project item

Because of the way `SplQueue` is implemented when you do a `dequeue` the
item is removed from within the queue, this will make TM to run the processors
for incoming and outcoming data only once.

Setting the queue in iterator mode and passing `SplDoublyLinkedList::IT_MODE_KEEP`
will prevent the processors to be removed which means them will run
for every project item.