Support gutenberg add posts to projects
Difference classic editor to gutenberg: Gutenberg save posts
asynchronously. That leads to the problem that the old way
(clicking on a button which holds a name) did not work.
Quick fix: When adding a post to a project make that
(by adding an input field with that name) available again.